third_party/blink/web_tests/fast/events/touch/touch-before-pressing-spin-button-expected.txt

Test if a spin-button works correctly after touch events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS numberInput.value is "2"
PASS successfullyParsed is true

TEST COMPLETE
